Clean up imports and rename deleteConversation to fix typo
